+++ src/components/Card/index.js
@@ -4,21 +4,29 @@ import { Thumbnail, Label } from 'native-base';
 import styles from './styles'
 const card = (props) => {
     return (
-        <View style={styles.container}>
-            <View>
-                <Thumbnail style={styles.thumbnailStyle} square large source={{ uri: props.picture }} />
-            </View>
-            <View style={styles.footerStyle}>
-                <Text numberOfLines={2} style={styles.nameStyle}>{props.name}</Text>
-                <View style={styles.popularityView}>
-                    <Text style={styles.popularityScore}>Rating</Text>
-                    <View style={styles.popularityLabel}>
-                        <Text style={styles.popularityScoreNumber}>{props.averageRating}</Text>
-                    </View>
+        <TouchableOpacity onPress={props.onPressCard}>
+            <View style={styles.container}>
+                <View>
+                    <Thumbnail style={styles.thumbnailStyle} square large source={{ uri: props.picture }} />
                 </View>
-                <Text style={styles.popularityScore}>{props.numberOfEpisodes} Eps - {props.minutesPerEpisode} Minutes</Text>
+                <View style={styles.footerStyle}>
+                    <Text numberOfLines={2} style={styles.nameStyle}>{props.name}</Text>
+                    <View style={styles.popularityView}>
+                        <Text style={styles.popularityScore}>Popularity</Text>
+                        <View style={styles.popularityLabel}>
+                            <Text style={styles.popularityScoreNumber}>{props.averageRating}</Text>
+                        </View>
+                    </View>
+                    {props.mangaTypeActive ?
+                        (
+                            <Text style={styles.popularityScore}>{props.numberOfEpisodes} Chapters - {props.minutesPerEpisode} {props.minutesPerEpisode === 1 ? 'Book' : 'Books'}</Text>
+                        )
+                        :
+                        <Text style={styles.popularityScore}>{props.numberOfEpisodes} Eps - {props.minutesPerEpisode} {props.minutesPerEpisode === 1 ? 'Minute' : 'Minutes'}</Text>
+                    }
+                </View>
             </View>
-        </View>
+        </TouchableOpacity>
     );
 }
 export default card