+++ src/components/Category/index.js
@@ -3,6 +3,7 @@ import React, { Component } from 'react';
 import { View, FlatList, TouchableHighlightBase, Text } from 'react-native';
 import { getDataByCategory } from '../../api/home'
 import { connect } from 'react-redux';
+import { withNavigationFocus } from 'react-navigation';
 //customs
 import styles from './styles';
 import Card from '../../components/Card'
@@ -11,7 +12,7 @@ class Category extends Component {
     loading: false,
     next_page_url: null,
     categoryList: [],
-    showCategoryName: false
+    showCategoryName: false,
   }
   componentDidMount = async () => {
     try {
@@ -31,35 +32,50 @@ class Category extends Component {
       console.log(err)
     }
   }
+  seeDetail = async index => {
+    try {
+      let { categoryList } = this.state;
+      console.log(categoryList[index],'categoryListsdklfjsdlkf')
+      let cardId = categoryList[index].id;
+      categoryList.id = cardId;
+      this.setState({ categoryList }, () => {
+        this.props.navigation.navigate('Detail', {
+          cardInformation: this.state.categoryList[index],
+        });
+      });
+    } catch (err) {
+      console.log(err);
+    }
+  }
   render() {
     let { showCategoryName, categoryList } = this.state;
     let { categoryName, type } = this.props;
     return (
-      <View style={styles.container}>
-        <FlatList
-          data={categoryList}
-          horizontal
-          showsHorizontalScrollIndicator={false}
-          renderItem={({ item, index }) => (
-            <View>
-              {showCategoryName ? (
-                this.props.showCategoryName
-              ) :
-                (null)}
-              <Card
-                averageRating={item.attributes.averageRating !== null ? item.attributes.averageRating : '--'}
-                picture={item.attributes.posterImage.original}
-                name={item.attributes.canonicalTitle}
-                numberOfEpisodes={item.attributes.episodeCount}
-                minutesPerEpisode={item.attributes.episodeLength}
-              />
-            </View>
-          )
-          }
-        />
-      </View>
+      showCategoryName ?
+        (<View style={styles.container}>
+          {showCategoryName ? (
+            this.props.showCategoryName
+          ) :
+            (null)}
+          <FlatList
+            data={categoryList}
+            horizontal
+            showsHorizontalScrollIndicator={false}
+            renderItem={({ item, index }) => (
+                <Card
+                  onPressCard={() => this.seeDetail(index)}
+                  averageRating={item.attributes.popularityRank !== null ? item.attributes.popularityRank : '--'}
+                  picture={item.attributes.posterImage.original}
+                  name={item.attributes.canonicalTitle}
+                  numberOfEpisodes={type === 'manga' ? item.attributes.chapterCount : item.attributes.episodeCount}
+                  minutesPerEpisode={type === 'manga' ? item.attributes.volumeCount : item.attributes.episodeLength}
+                  mangaTypeActive={type === 'manga' ? true : false}
+                />
+            )}
+          />
+        </View>) : null
     );
   }
 }
 
-export default Category
+export default withNavigationFocus(Category)