+++ src/components/FooterScrollableTab/index.js
@@ -2,25 +2,23 @@ import React from 'react';
 import { Container, Tab, Tabs, ScrollableTab } from 'native-base';
 import styles from './styles';
 import { View, Text } from 'react-native';
-function tabBar(props) {
+function tabBarScrollable(props) {
     return (
         <Container style={styles.backgroundContainer}>
             <Tabs
                 initialPage={0}
                 style={{ borderWidth: 0}}
-                // onChangeTab={info => props.onChangeTab(info)}
                 tabBarUnderlineStyle={styles.borderBottomTabColor}
-                // renderTabBar={() => (
-                //     <ScrollableTab style={{ borderWidth: 0}} tabsContainerStyle={styles.colorScroll} />
-                // )}
-                locked
-                >
+                renderTabBar={() => (
+                    <ScrollableTab style={{ borderWidth: 0 }} tabsContainerStyle={styles.colorScroll} />
+                )}
+            >
                 <Tab
                     textStyle={styles.textStyleTab}
                     activeTextStyle={styles.activeTextStyleTab}
                     activeTabStyle={styles.activeTabColorBackground}
                     tabStyle={styles.TabColorBackground}
-                    heading={'Anime'}
+                    heading={'Synopsis'}
                 >
                     {props.contentTab1}
                 </Tab>
@@ -29,11 +27,11 @@ function tabBar(props) {
                     activeTextStyle={styles.activeTextStyleTab}
                     activeTabStyle={styles.activeTabColorBackground}
                     tabStyle={styles.TabColorBackground}
-                    heading={'Manga'}>
+                    heading={'Technical Details'}>
                     {props.contentTab2}
                 </Tab>
             </Tabs>
         </Container>
     );
 }
-export default tabBar;
+export default tabBarScrollable;