+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-## [ADD] 2020-02-12 21:00
+## [ADD] 2020-02-12 21:00PM
 - Initialize app
 
-## [ADD] 2020-02-13 12:00
+## [ADD] 2020-02-13 12:00PM
 - Created component for Anime Cards
 - Created component for TabBarScrollable
 - Created component for Categories
 - Added FlatList for categories and anime cards
+
+## [ADD] 2020-02-13 10:45PM
+- Created component for Detail screen
+- Fetching data from API to Detail Screen